refactor(root): adds script for overwritting manifests
abstracts out dynamic configs from gatsby-config, adds stock maniftes files, updates ci config
