+++ utils/manifest.webapp
@@ -1,0 +1,14 @@
+{
+  "version": "1.0",
+  "name": "codybrunner.rocks",
+  "description": "Personal website, tech blog, & portfolio for Cody Brunner",
+  "icons": {
+    "60": "firefox_app_60x60.png",
+    "128": "firefox_app_128x128.png",
+    "512": "firefox_app_512x512.png"
+  },
+  "developer": {
+    "name": "Cody Brunner",
+    "url": "https://codybrunner.rocks"
+  }
+}
